refactor child invest repository package
