+++ isol/server-common/src/main/java/shinhan/server_common/domain/invest/service/InvestProposalService.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shinhan.server_common.domain.invest.investEntity.InvestProposalResponse;
-import shinhan.server_common.domain.invest.repository.InvestProposalResponseRepository;
+import shinhan.server_common.domain.invest.investRepository.InvestProposalResponseRepository;
 import shinhan.server_common.domain.invest.repository.CorpCodeRepository;
 
 @Service
