+++ isol/server-common/src/main/java/shinhan/server_common/domain/invest/service/InvestService.java
@@ -19,8 +19,8 @@ import shinhan.server_common.domain.invest.investEntity.Portfolio;
 import shinhan.server_common.domain.invest.investEntity.StockHistory;
 import shinhan.server_common.domain.invest.repository.CorpCodeRepository;
 import shinhan.server_common.domain.invest.dto.StockFindCurrentResponse;
-import shinhan.server_common.domain.invest.repository.PortfolioRepository;
-import shinhan.server_common.domain.invest.repository.StockHistoryRepository;
+import shinhan.server_common.domain.invest.investRepository.PortfolioRepository;
+import shinhan.server_common.domain.invest.investRepository.StockHistoryRepository;
 import shinhan.server_common.domain.invest.repository.StockRepository;
 import shinhan.server_common.global.exception.CustomException;
 import shinhan.server_common.global.utils.account.AccountUtils;
